refactor: Extract drag handle
